# use batched gemm to speed up GPU and lower the JIT time