--- a/Project.toml
+++ b/Project.toml
@@ -4,6 +4,7 @@ authors = ["Andreas Peter <andreas.peter.ch@gmail.com>"]
 version = "0.3.0"
 
 [deps]
+BatchedRoutines = "a9ab73d0-e05c-5df1-8fde-d6a4645b8d8e"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 IRTools = "7869d1d1-7146-5819-86e3-90919afe41df"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"