--- a/Project.toml
+++ b/Project.toml
@@ -17,6 +17,7 @@ TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
+BatchedRoutines = "~0.2.0"
 julia = "1"
 
 [extras]