--- a/src/EinRule.jl
+++ b/src/EinRule.jl
@@ -110,6 +110,7 @@ const einsum_rules = [
 @doc raw"
     match_rule(ixs, iy)
     match_rule(code::EinCode{ixs, iy})
+    match_rule(code::NestedEinCode)
 
 go through all operations specified in the `einsum_rules`-vector and return
 the first `T` for which `match_rule(T, ixs, iy)` returns true.
@@ -123,3 +124,9 @@ function match_rule(ixs, iy)
 end
 
 match_rule(code::EinCode{ixs, iy}) where {ixs, iy} = match_rule(ixs, iy)
+
+function match_rule(code::NestedEinsumStable)
+    return (match_rule(code.eins), match_rule.(code.args))
+end
+
+match_rule(code::Int64) = code