--- a/src/utils.jl
+++ b/src/utils.jl
@@ -59,14 +59,7 @@ julia> allunique((1,2,3,1))
 false
 ```
 """
-function allunique(ix::NTuple{N,T}) where {N,T}
-    for i=1:N
-        for j=i+1:N
-            @inbounds ix[i] == ix[j] && return false
-        end
-    end
-    return true
-end
+allunique(ix::NTuple) = all(i -> count(==(i), ix) == 1, ix)
 
 function conditioned_permutedims(A::AbstractArray{T,N}, pA) where {T,N}
     any(i-> (@inbounds pA[i]!=i), 1:N) ? permutedims(A, pA) : A