Resolving #74 .

It can be merged after @Roger-luo tag the new release of BatchedRoutines.

Note: this is adapted from #2