@GiggleLiu 2019-09-08
# Benchmark
## Rule change
#### master branch
```
julia> OMEinsum.match_rule(ein"ijl,jkl->ikl")
OMEinsum.DefaultRule()

julia> OMEinsum.match_rule(ein"ijl,jkl->ik")
OMEinsum.PairWise()
```

#### new
```julia
julia> OMEinsum.match_rule(ein"ijl,jkl->ikl")
OMEinsum.BatchedContract()

julia> OMEinsum.match_rule(ein"ijl,jkl->ik")
OMEinsum.BatchedContract()
```

## benchmark - CPU
#### master
```julia
T = randn(100,100,1000);
julia> @benchmark ein"ijk,jlk->ilk"(T, T)
BenchmarkTools.Trial: 
  memory estimate:  76.31 MiB
  allocs estimate:  240
  --------------
  minimum time:     89.986 ms (9.70% GC)
  median time:      96.308 ms (8.90% GC)
  mean time:        109.259 ms (12.13% GC)
  maximum time:     207.974 ms (51.01% GC)
  --------------
  samples:          47
  evals/sample:     1

julia> @benchmark ein"ijk,jlk->il"(T, T)
BenchmarkTools.Trial: 
  memory estimate:  76.38 MiB
  allocs estimate:  238
  --------------
  minimum time:     133.128 ms (13.56% GC)
  median time:      152.739 ms (5.61% GC)
  mean time:        156.250 ms (7.65% GC)
  maximum time:     223.539 ms (45.41% GC)
  --------------
  samples:          32
  evals/sample:     1
```

#### new
```julia
julia> @benchmark ein"ijk,jlk->ilk"(T, T)
BenchmarkTools.Trial: 
  memory estimate:  76.30 MiB
  allocs estimate:  25
  --------------
  minimum time:     1.580 s (1.38% GC)
  median time:      1.611 s (2.59% GC)
  mean time:        1.606 s (2.59% GC)
  maximum time:     1.623 s (4.04% GC)
  --------------
  samples:          4
  evals/sample:     1

julia> @benchmark ein"ijk,jlk->il"(T, T)
BenchmarkTools.Trial: 
  memory estimate:  79.11 KiB
  allocs estimate:  22
  --------------
  minimum time:     81.214 ms (0.00% GC)
  median time:      85.310 ms (0.00% GC)
  mean time:        86.902 ms (0.00% GC)
  maximum time:     130.337 ms (0.00% GC)
  --------------
  samples:          58
  evals/sample:     1
```

## benchmark - GPU
#### master branch
```julia
julia> @benchmark (CuArrays.@sync ein"ijk,jlk->ilk"(T, T))
BenchmarkTools.Trial: 
  memory estimate:  10.50 KiB
  allocs estimate:  205
  --------------
  minimum time:     135.596 ms (0.00% GC)
  median time:      138.661 ms (0.00% GC)
  mean time:        139.749 ms (0.29% GC)
  maximum time:     153.708 ms (2.63% GC)
  --------------
  samples:          36
  evals/sample:     1

julia> @benchmark (CuArrays.@sync ein"ijk,jlk->il"(T, T))
BenchmarkTools.Trial: 
  memory estimate:  9.45 KiB
  allocs estimate:  185
  --------------
  minimum time:     1.873 s (0.00% GC)
  median time:      1.878 s (0.00% GC)
  mean time:        1.896 s (0.00% GC)
  maximum time:     1.938 s (0.00% GC)
  --------------
  samples:          3
  evals/sample:     1
```
#### new
```julia
julia> @benchmark (CuArrays.@sync ein"ijk,jlk->ilk"(T, T))
BenchmarkTools.Trial: 
  memory estimate:  15.48 KiB
  allocs estimate:  269
  --------------
  minimum time:     130.312 ms (0.00% GC)
  median time:      130.936 ms (0.00% GC)
  mean time:        132.623 ms (0.20% GC)
  maximum time:     143.398 ms (1.55% GC)
  --------------
  samples:          38
  evals/sample:     1

julia> @benchmark (CuArrays.@sync ein"ijk,jlk->il"(T, T))
BenchmarkTools.Trial: 
  memory estimate:  17.36 KiB
  allocs estimate:  320
  --------------
  minimum time:     131.481 ms (0.00% GC)
  median time:      134.130 ms (0.00% GC)
  mean time:        141.025 ms (0.19% GC)
  maximum time:     180.895 ms (1.67% GC)
  --------------
  samples:          36
  evals/sample:     1
```

## What can be improved
* the index information can be obtained statially by seperating function `analyse_gemm`, but I didn't do that to avoid potential compiling overhead.
* the performance of CPU version pairwise contraction falls back, but @Roger-luo said the performance is much better in the next version by approximately a factor of 2.