@Roger-luo 2019-09-10
I removed the conditional dependency hack in BatchedRoutines, use CuBatchedRoutines instead.

As rust-cargo says:

> the canonical way of implementing conditional dependencies is to have multiple packages