@Roger-luo 2019-09-11
yes, you can wrap up that routine in cublas yourself, but the problem is CuBatchedRoutines provides a unified API simply load `CuBatchedRoutines` with `Require` will do the job (in principal you don't need to write anything else). I can register `CuBatchedRoutines` a bit later.

And there're a few other routines which is not in cublas that I think you will also want to dispatch, e.g `batched_tr` etc.