@Roger-luo 2019-09-13
use the lastest release, I removed CUDA related hack in the package. It had some issue on CI before.