@GiggleLiu 2019-09-13
> use the lastest release, I removed CUDA related hack in the package. It had some issue on CI before.

:+1: Thanks, having changed the compat, now the tests pass!